Phase K Wave 21: 0x002D ACTION2 round-trip + sizeof(u_long) tightening
Adds SectorAction2Tests.Action2_NoOpSubActionAndEmptyName_DoesNotBreak
Connection_RequestTimeStillRoundTrips — sends min 14-byte ActionPacket2
(GameID=0 BE, Action=23 BE, string_len=0 LE, OptionalVar=0 BE) and
asserts survival via REQUEST_TIME -> CLIENT_SET_TIME with sentinel tick.

Action=23 is the literal commented-out no-op (`//keep trading???`
server/src/PlayerConnection.cpp:4104), same branch Wave 13 used for
0x002C ACTION. string_len=0 deliberately exercises OptionalVar offset
arithmetic with zero string slack — worst case for the read-width bug
class.

Server tightening shipped in same wave (preservation-aligned per
CLAUDE.md): fixed the one `*(u_long *)` cast Wave 12's sweep missed at
server/src/PlayerConnection.cpp:4265 (HandleAction2 OptionalVar read).
Wave 12 grepped `*((long *)` and `*((unsigned long *)` but the POSIX
`u_long` typedef variant was not matched. On Win32 sizeof(u_long)==4
so the cast was accidentally correct; on Linux x86_64 it reads 4 bytes
past the wire's int32 OptionalVar field. uint32_t* restores 4B parity
with retail. Verified the only site via grep.

7 regression classes covered: (1) Wave-12-class read-width on Linux,
(2) Player->Action2 dispatch, (3) Action2->Action ntohl conversion,
(4) GetGameIDFromName("") empty-string safety, (5) ObjectManager
negative-ID safety, (6) HandleAction case-23 chain safety post-
conversion, (7) connection survival.

Pool[16] cli_test18 dedicated. Coverage 31/207 -> 32/207 (15.5%).
Isolated test passes in 7s; full integration suite 52/52 in 1m54s
from a freshly-recycled stack.

+++ server/src/PlayerConnection.cpp
@@ -4262,7 +4262,11 @@ void Player::HandleAction2(unsigned char *data)
 	converted.GameID = ntohl(myAction2->GameID);
 	converted.Action = ntohl(myAction2->Action);
 	converted.Target = g_PlayerMgr->GetGameIDFromName(myAction2->string);
-	converted.OptionalVar = ntohl(*(u_long *)(myAction2->string+myAction2->string_len));
+	// uint32_t* (not u_long*): sizeof(u_long)==8 on Linux x86_64 would read 4 bytes past
+	// the wire's int32 OptionalVar field. Same bug class fixed sweep-wide in Wave 12;
+	// missed there because Wave 12's grep covered `unsigned long` but not the `u_long`
+	// POSIX typedef. See plans/99-decisions-log.md 2026-05-25 Wave 12.
+	converted.OptionalVar = ntohl(*(uint32_t *)(myAction2->string+myAction2->string_len));
 
 	HandleAction((unsigned char *)&converted);
 }
